added userType
