+++ userType.ts
@@ -1,0 +1,13 @@
+type User = {
+  firstName: string;
+  lastName: string;
+  address: string;
+  city: string;
+  country: string;
+  emailAddress: string;
+  zipcode: number;
+  cart: string[];
+  owned: string[];
+  wishlist: string[];
+};
+export default User;
